backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
slowapi==0.1.9
python-multipart==0.0.9
sqlalchemy==2.0.36
pymysql==1.1.1
requests==2.32.3
scikit-learn==1.6.1
numpy==2.1.3
python-dotenv==1.0.1
